requirements.txt
httpx
bilibili-api-python
pydantic~=2.10.5
ffmpeg-python~=0.2.0
numpy
soundfile
openai-whisper
loguru